refactor(parser): split into parsers/ package (Python verbatim) (#42)
Phase 0 of multi-language AST support: split the 645-line parser.py into a
graphlm/parsers/ package so future languages plug in without touching the
Python resolver. Pure refactor — zero behavior change.

- parsers/base.py: registry-driven _TreeSitterBackend (single _backend
  singleton; lazy per-language grammar import via _GRAMMARS/_GrammarSpec;
  _GrammarUnavailable degrades a missing grammar to zero edges for that
  language, never escapes build_dependency_graph/parse_file), ParsedFile,
  shared helpers, group-by-language dispatch, detect_import_cycles.
- parsers/python.py: Python resolver moved verbatim (queries, import
  extraction, _source_roots + src-layout resolution — #19 behavior
  unchanged), registered for "python".
- parser.py: thin backwards-compat shim re-exporting the public contract
  (incl. _source_roots).
- New test asserts the never-escapes invariant with real teeth: a fake
  language registered in BOTH _GRAMMARS and the extension map yields [] (not
  None) for that language while Python edges stay intact.
- CLAUDE.md: parser.py path references updated to the new module locations.

Verified: 396 tests pass (395 unmodified + 1 new), mypy clean, all modules
under 600 lines, deterministic-edge count on the pristine HEAD tree unchanged
at 51 (verbatim-move proof).
