+++ CLAUDE.md
@@ -30,9 +30,9 @@ The *analysis* pipeline lives in `graphlm/__init__.py::generate_graph()` — rea
 
 `--dry-run` skips both LLM calls. It does *not* model "all scanned files" — it selects `scan.file_fragments[:max_pass2_files]` (default cap 80), builds a `CodebaseGraph` carrying the AST edges + cycles + a dry-run note, and the CLI prints token estimates *and* graph-section counts. Tokens are estimated by a crude heuristic — `estimate_tokens` = `UTF-8 bytes * 2 // 5` (≈2.5 bytes/token). It is defined once in `scanner.py` and re-exported by `context.py`. The ratio is calibrated to over-estimate: real content measured ~2.83 bytes/token against the served model, so the old `// 4` (4 bytes/token) *under*-counted by ~28% and let graphlm pack prompts the model then rejected or timed out on — the `* 2 // 5` divisor sits safely above the real count (#17).
 
-**AST parser is ground truth, not a replacement for the LLM** (`parser.py`, on by default; `--no-ast` disables). `build_dependency_graph()` extracts Python import edges with Tree-sitter and resolves them *only against files that exist in the scan* (stdlib/third-party/missing modules are dropped — resolution is against the `known_files` set, independent of `project_dir`). Because edges resolve only between *scanned* files, the scanner's `_rank_file` ranks source code (`_SOURCE_EXTS`) above non-source text so a doc-heavy repo can't crowd source out of the `max_files` cap and starve the graph (#19). Resolution handles **src-layout** projects: `_source_roots()` accepts a source-root prefix only when its final segment is a conventional root name (`_SOURCE_ROOT_NAMES` = `src`/`lib`/`python`) that directly contains a package, and each import candidate is tried under every accepted root (root `""` first, so root-layout projects are unchanged) — without this, a `from mypkg.x import Y` in a `src/mypkg/` project resolved to nothing and the whole AST graph came back empty (#19). The allowlist is deliberate: deriving a root from *any* directory containing an `__init__.py` let a `tests/stub/requests/__init__.py` shadow make third-party `import requests` resolve to a project-internal file — a false edge in the "do-not-contradict" ground-truth table, worse than a missing one. `project_dir` is *optional*: given, `_source_bytes` reads the untruncated/unredacted file off disk; omitted, it falls back to the (truncated) `frag.content`. Those `deterministic_edges` are (a) injected into the pass-2 prompt as a "do not contradict" table, and (b) used for cycle detection instead of the LLM's edges. The two uses diverge under a tight `max_context`: the **prompt** copy of the table is capped at a share of the budget (`EDGE_SHARE`, `context.py`) and, when capped, its framing flips to "not exhaustive — infer the rest" so a partial list isn't presented as complete; the **full** list always reaches `graph.deterministic_edges` and cycle detection. So the LLM's `import_edges` and the AST's `deterministic_edges` are separate fields on `CodebaseGraph` and can differ. Only Python is fully implemented; JS/TS are recognized by extension but return empty `ParsedFile`.
+**AST parser is ground truth, not a replacement for the LLM** (the `graphlm/parsers/` package — `base.py` is the registry-driven Tree-sitter backend + language-agnostic dispatch, `python.py` is the Python resolver; `graphlm/parser.py` is a thin backwards-compat shim re-exporting the public contract, incl. `_source_roots`; on by default, `--no-ast` disables). `build_dependency_graph()` extracts Python import edges with Tree-sitter and resolves them *only against files that exist in the scan* (stdlib/third-party/missing modules are dropped — resolution is against the `known_files` set, independent of `project_dir`). Because edges resolve only between *scanned* files, the scanner's `_rank_file` ranks source code (`_SOURCE_EXTS`) above non-source text so a doc-heavy repo can't crowd source out of the `max_files` cap and starve the graph (#19). Resolution handles **src-layout** projects: `_source_roots()` accepts a source-root prefix only when its final segment is a conventional root name (`_SOURCE_ROOT_NAMES` = `src`/`lib`/`python`) that directly contains a package, and each import candidate is tried under every accepted root (root `""` first, so root-layout projects are unchanged) — without this, a `from mypkg.x import Y` in a `src/mypkg/` project resolved to nothing and the whole AST graph came back empty (#19). The allowlist is deliberate: deriving a root from *any* directory containing an `__init__.py` let a `tests/stub/requests/__init__.py` shadow make third-party `import requests` resolve to a project-internal file — a false edge in the "do-not-contradict" ground-truth table, worse than a missing one. `project_dir` is *optional*: given, `_source_bytes` reads the untruncated/unredacted file off disk; omitted, it falls back to the (truncated) `frag.content`. Those `deterministic_edges` are (a) injected into the pass-2 prompt as a "do not contradict" table, and (b) used for cycle detection instead of the LLM's edges. The two uses diverge under a tight `max_context`: the **prompt** copy of the table is capped at a share of the budget (`EDGE_SHARE`, `context.py`) and, when capped, its framing flips to "not exhaustive — infer the rest" so a partial list isn't presented as complete; the **full** list always reaches `graph.deterministic_edges` and cycle detection. So the LLM's `import_edges` and the AST's `deterministic_edges` are separate fields on `CodebaseGraph` and can differ. Only Python is fully implemented; JS/TS are recognized by extension but return empty `ParsedFile`.
 
-**Cycle detection** (`cycles.py`) runs Tarjan SCC over the AST edges (falling back to LLM edges if AST is off), scoring each cycle `log10(total_lines) * cycle_length`. Note the SLOC term is a *physical* line count (`frag.content.count("\n") + 1`, including blanks/comments) over the scanned fragment — which is truncated at `max_file_chars` (default 4000) — so it under-counts large files; it is not true source-lines-of-code. Also: `parser.py` *also* has its own `detect_import_cycles()` (returns bare node lists, no scores) — the live path uses `cycles.detect_cycles()`; don't confuse the two.
+**Cycle detection** (`cycles.py`) runs Tarjan SCC over the AST edges (falling back to LLM edges if AST is off), scoring each cycle `log10(total_lines) * cycle_length`. Note the SLOC term is a *physical* line count (`frag.content.count("\n") + 1`, including blanks/comments) over the scanned fragment — which is truncated at `max_file_chars` (default 4000) — so it under-counts large files; it is not true source-lines-of-code. Also: `graphlm/parsers/base.py` *also* has its own `detect_import_cycles()` (returns bare node lists, no scores; re-exported via the `graphlm/parser.py` shim) — the live path uses `cycles.detect_cycles()`; don't confuse the two.
 
 **Data models** (`models.py`) — `CodebaseGraph` is the Pydantic schema the LLM must emit and the single source of truth for output shape. `Cycle` is a frozen dataclass (not Pydantic). The pass-2 prompt in `context.py` hand-writes the JSON schema description for the LLM; **if you add/rename a field on `CodebaseGraph`, update that prompt text and `render.py` in the same change** — they are not auto-derived from the model. **Exception — `meta` (`GraphMeta`):** the provenance stamp is *not* LLM-emitted (like `directory_tree`/`deterministic_edges`, it's filled locally in `generate_graph`), so it is deliberately **not** in the pass-2 instruction block — don't add it. Its `schema_version` (`GRAPH_META_SCHEMA_VERSION`) is a versioned *input* contract: graphlm reads its own prior `GRAPH.json` for the fast-follow diff, so bump it if the meta shape changes (see the self-refreshing-graph ADR in `DECISIONS.md`).
